Add back deleted, deprecated accumulate exercise
When migrating the data from v1 to v2, we will lose all the solutions submitted to
deprecated exercises, unless they exist in the track configuration.

This adds the deprecated parallel-letter-frequency exercise back, leaving it deprecated so that it
doesn't actively show up as part of the track.

We can delete it from the track again after we launch v2, if we want to.