Copy Ruby Track's core progression (#1633)
This is to get you all started with the rearrangement of the track, as proposed in #1614
Please note that I don't know Python; this was done based on the changes I made to the Ruby track. 
Hope it helps!
 
1. Copy Ruby's core exercise progression as a first step to help decrease the mentoring workload.
2. For easy reference, I put all the core exercises at the top of the file. 
3. Removed Leap as a core exercises, and choose other cores to unlock the side exercises that used to depend on Leap
4. Raindrops was marked 'deprecated', but is part of the Ruby core; I reinstated it for now.
5. In Ruby Difference of Squares and Grain are still cores, but will be changed to sides. As Python didn't have them as core, I left it like that. 
6. In Ruby, Robot Name is a core, but it is not an exercise in Python. To avoid extra work, I added Word Count instead. https://github.com/exercism/python/pull/1633/commits/00d5afa8fe8de8e1c0fe526797d66bd27ab47162
7. Extra cores that are in Python, not in Ruby:
Bob, Allergies, Sum of Multiples, Kindergarten, Grade School, Saddle Points, BinarySearch, List ops, Bookstore, Markdown, Rest-api.
I added commits for each of them, so that you can easily revert them. My changes are based on either considerations I made to the Ruby track, or a rough estimation in regard to the changes in this PR. 
- [ ] Kept as core: Kindergarten Garden; Grade School; Book Store; Markdown.
- [ ] Changed into side: Bob, Allergies, Sum of Multiples, Saddle Points, Binary Search, List Ops; Rest-API.


*Closes #1614*

Maintainers To do :
- [ ]  Squash PR before merging
- [ ]  Ping Jeremy and ask to un-deprecate Python/Raindrops