Remove duplicate instructions header
There is both a first level and a second-level header with the text "Instructions". This is not a big problem in the exercise description, but in the editor, the first task is shown as "Instructions", which is not right. I've removed the duplicate header so that Task 1 will be shown correctly as "Tracking Poker Rounds".