Update to the test_is_blackjack function
Solution/Modification: Added two extra test cases to the data array in the test_is_blackjack function.

Problem: Before, one could conceivably pass the test_is_blackjack test by calling the value_of_card function for each card, add the results together, and then test if the value was 11 (as an Ace in the value_of_card function was fixed as 1).  However, this alone does not necessarily indicate that blackjack has been achieved, as any two cards that sum to 11 (such as 4 and 7, or 9 and 2) would also pass the test, even though they shouldn't.  For a blackjack to occur, one must also check that one of the cards is an Ace (or, alternatively, that one of the cards has a value of 10)