Remove lasagna from the constant name (EXPECTED_BAKE_TIME)
The first part of the exercise consist on defining a constant. In the instructions there is an example where the the constant is called as follows:
>>> lasagna.EXPECTED_BAKE_TIME
I believe that the constant was probably an atribute from the lasagna class in an earlier version of the problem, and now it just makes it difficult to anderstand what "lasagna" means in the exercise.