Fail closed on pause contributions workflow
In the original workflow, I had `catch(err) => false`.

But during code review the question was raised:
What happens if we get rate limited and all the requests from the API fail?
(E.g. if we do a script and automatically create PRs across all of Exercism, which is totally a thing).

We were like: ooh, that would suck, wouldn’t it? It would be better if we occasionally had to deal with manually closing a PR.
But here’s the kicker. The API response has no body. It’s either 204 or 404, where 404 is perceived as… an error.

So I changed it, and (importantly) forgot to test the script one final time.

So here is a version that will actually work. (I tested.)